requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
sqlite3-binary
